\name{filling.d}
\alias{filling.d}

\title{Fill in missing values}
\description{
Filling in missing values of a daily timeSeries object
}
\usage{
filling.d(x)
}
\arguments{
  \item{x}{daily timeSeries object}
}
\details{See McLeod, Hipel and Camacho(1983) for further details.}
\value{daily timeSeries object}
\references{
McLeod, A.I., Hipel, K.W. and Camacho, F. (1983), 
Trend assessment of water quality time series, Water Resources Bulletin, 
19 537-547.
}
\author{
Hyukjun Gweon and A.I. McLeod
}
