\name{sltl-package}
\alias{sltl-package}
\docType{package}
\title{
Time series decompoisiton using loess and harmonic regression
}
\description{      
Decompose daily or monthly timeSeries object into seasonal and other components.
Loess is used for long-term components while a seasonal pattern is estimated by harmonic regression.
If an unevenly-spaced timeSeries object is entered, missing values can be estimaed based on 
the seasonal adjustment algorithm.
}
\details{
\tabular{ll}{
Package: \tab sltl\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-18\cr
License: \tab GPL (>= 2.10)\cr
}
}
\author{
Hyukjun Gweon and A. I. McLeod
 <hgweon@uwo.ca>
}
\references{

McLeod, A.I., Hipel, K.W. and Camacho, F. (1983), 
Trend assessment of water quality time series, Water Resources Bulletin, 
19 537-547.

McLeod, A. I. and  Hyukjun Gweon.  (2013). 
Optimal Deseasonalization for Geophysical Time Series. 
Journal of Environmental Statistics. Vol. 4, Issue 11
}

\keyword{ package }

